requirements.txt
streamlit==0.86.0
pandas==0.24.2
joblib==0.15.1
shap==0.39.0
Pillow==8.3.0
scikit_learn==0.23.1
